DoubleField renamed to the more appropriate DecimalField BigDecimal support continued (as in, things I missed)
git-svn-id: svn://svn.hafro.is/jminor/trunk@17513 51816742-9f83-410d-85ed-e7bcf7042953

+++ swing/common-ui/src/test/java/org/jminor/swing/common/ui/input/DoubleInputProviderTest.java
@@ -3,7 +3,7 @@
  */
 package org.jminor.swing.common.ui.input;
 
-import org.jminor.swing.common.ui.textfield.DoubleField;
+import org.jminor.swing.common.ui.textfield.DecimalField;
 
 import org.junit.jupiter.api.Test;
 
@@ -26,13 +26,13 @@ public class DoubleInputProviderTest {
     final DoubleInputProvider provider = new DoubleInputProvider(null);
     assertNull(provider.getValue());
 
-    ((DoubleField) provider.getInputComponent()).setSeparators('.', ',');
+    ((DecimalField) provider.getInputComponent()).setSeparators('.', ',');
     provider.getInputComponent().setText("15.5");
     assertEquals(Double.valueOf(15.5), provider.getValue());
     provider.getInputComponent().setText("15,6");
     assertEquals(Double.valueOf(15.5), provider.getValue());
 
-    ((DoubleField) provider.getInputComponent()).setSeparators(',', '.');
+    ((DecimalField) provider.getInputComponent()).setSeparators(',', '.');
     provider.getInputComponent().setText("15.7");
     assertEquals(Double.valueOf(15.5), provider.getValue());
     provider.getInputComponent().setText("15,7");
